// src/range/core/type.d.ts
export type TLocale = "fa" | "en";
// ...
export interface IDate {
    from: number;
    to: number;
}